requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
frozenlist==1.4.1
idna==3.10
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==75.1.0
six==1.16.0
typing_extensions==4.12.2
yarl==1.15.2
